requirements.txt
numpy==1.24.1
streamlit==1.17.0
tensorflow-gpu==2.10.1
